+++ CHANGELOG.md
@@ -7,6 +7,7 @@
   - datacenter support for API300
   - fabric new resource support for API300 Synergy
   - san_manager support for API300
+  - firmware support for API300
 
 ### 1.1.0
   - Add support for client ENV variables